tests: 64 bit compatibility
Fixed compilation errors. Mostly DEBUG/printf formatting and void pointer casting.

Other changes are:
* net/gnrc_sixlowpan_frag_*: Generalized packet size calculation
* cpu/native_backtrace: Reduced required backtrace size to 3 for 64-bit
* periph/flashpage: Simplified test
* unittests/tests-pktbuf: Generalized alignment
* sys/architecture: Extended test for 64-bit

+++ tests/drivers/at86rf215/main.c
@@ -191,7 +191,7 @@ int netdev_ieee802154_minimal_init_devs(netdev_event_cb_t cb) {
         at86rf215_t *at86rf215_subghz = NULL;
         at86rf215_t *at86rf215_24ghz = NULL;
 
-        printf("%d out of %d\n", i + 1, AT86RF215_NUM);
+        printf("%d out of %u\n", i + 1, (unsigned)AT86RF215_NUM);
 
         if (IS_USED(MODULE_AT86RF215_SUBGHZ)) {
             puts("Sub-GHz");
